requirements.txt
APScheduler==3.6.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
python-firebase @ git+https://github.com/ozgur/python-firebase@0d79d7609844569ea1cec4ac71cb9038e834c355
python-telegram-bot==13.5
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tornado==6.4
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
python-firebase~=1.2.1